// shared/enums/DefensiveScheme.ts
export enum DefensiveSchemeEnum {
    FOUR_THREE = "FOUR_THREE",
    THREE_FOUR = "THREE_FOUR",
    HYBRID = "HYBRID",
    NICKEL_BASED = "NICKEL_BASED",
    COVER_2 = "COVER_2",
    COVER_3 = "COVER_3",
    MULTIPLE = "MULTIPLE",
    MAN_BLITZ_HEAVY = "MAN_BLITZ_HEAVY",
    ZONE_HEAVY = "ZONE_HEAVY",
}