// shared/enums/OffensiveScheme.ts
export enum OffensiveSchemeEnum {
    WEST_COAST = "WEST_COAST",
    AIR_RAID = "AIR_RAID",
    PRO_STYLE = "PRO_STYLE",
    WILDCAT = "WILDCAT",
    SPREAD = "SPREAD",
    POWER_RUN = "POWER_RUN",
    ZONE_RUN = "ZONE_RUN",
    BALANCED = "BALANCED",
    PISTOL = "PISTOL",
    RUN_AND_SHOOT = "RUN_AND_SHOOT"
}